+++ Ecoquest/sprites.c
@@ -1,7 +1,7 @@
 #include "sprites.h"
 #include <stdio.h>
 #include <stdlib.h>
-#include <string.h>
+// string.h removido pois não era usado diretamente aqui, exceto se necessário para alguma macro interna
 
 // ========== SPRITES ESTATICOS ==========
 
@@ -18,19 +18,20 @@ Sprite* criar_sprite(const char* caminho_arquivo) {
         sprite->carregado = false;
         sprite->largura = 0;
         sprite->altura = 0;
-    } else {
+    }
+    else {
         sprite->carregado = true;
         sprite->largura = (float)al_get_bitmap_width(sprite->imagem);
         sprite->altura = (float)al_get_bitmap_height(sprite->imagem);
     }
-    
+
     sprite->caminho_arquivo = caminho_arquivo;
     return sprite;
 }
 
 void desenhar_sprite(const Sprite* sprite, float x, float y, float escala) {
     if (!sprite || !sprite->carregado) return;
-    
+
     al_draw_scaled_bitmap(
         sprite->imagem,
         0, 0, sprite->largura, sprite->altura,
@@ -42,16 +43,16 @@ void desenhar_sprite(const Sprite* sprite, float x, float y, float escala) {
     );
 }
 
-void desenhar_sprite_camera(const Sprite* sprite, float x, float y, 
-                           float camera_x, float camera_y, 
-                           float escala, float zoom) {
+void desenhar_sprite_camera(const Sprite* sprite, float x, float y,
+    float camera_x, float camera_y,
+    float escala, float zoom) {
     if (!sprite || !sprite->carregado) return;
-    
+
     float x_tela = (x - camera_x) * zoom;
     float y_tela = (y - camera_y) * zoom;
     float largura_escalada = sprite->largura * escala * zoom;
     float altura_escalada = sprite->altura * escala * zoom;
-    
+
     al_draw_scaled_bitmap(
         sprite->imagem,
         0, 0, sprite->largura, sprite->altura,
@@ -63,10 +64,10 @@ void desenhar_sprite_camera(const Sprite* sprite, float x, float y,
     );
 }
 
-void desenhar_sprite_rotacionado(const Sprite* sprite, float x, float y, 
-                                float angulo, float escala) {
+void desenhar_sprite_rotacionado(const Sprite* sprite, float x, float y,
+    float angulo, float escala) {
     if (!sprite || !sprite->carregado) return;
-    
+
     al_draw_scaled_rotated_bitmap(
         sprite->imagem,
         sprite->largura / 2, sprite->altura / 2,
@@ -79,7 +80,7 @@ void desenhar_sprite_rotacionado(const Sprite* sprite, float x, float y,
 
 void destruir_sprite(Sprite* sprite) {
     if (!sprite) return;
-    
+
     if (sprite->imagem) {
         al_destroy_bitmap(sprite->imagem);
     }
@@ -88,10 +89,10 @@ void destruir_sprite(Sprite* sprite) {
 
 // ========== SPRITES ANIMADOS (SPRITESHEET) ==========
 
-SpriteAnimado* criar_sprite_animado(const char* caminho_arquivo, 
-                                   int frames_total, 
-                                   int frames_por_linha,
-                                   float tempo_por_frame) {
+SpriteAnimado* criar_sprite_animado(const char* caminho_arquivo,
+    int frames_total,
+    int frames_por_linha,
+    float tempo_por_frame) {
     SpriteAnimado* sprite = (SpriteAnimado*)malloc(sizeof(SpriteAnimado));
     if (!sprite) {
         fprintf(stderr, "Erro ao alocar memoria para sprite animado\n");
@@ -113,30 +114,31 @@ SpriteAnimado* criar_sprite_animado(const char* caminho_arquivo,
     sprite->tempo_por_frame = tempo_por_frame;
     sprite->tempo_acumulado = 0.0f;
     sprite->loop = true;
-    
+
     float largura_sheet = (float)al_get_bitmap_width(sprite->spritesheet);
     float altura_sheet = (float)al_get_bitmap_height(sprite->spritesheet);
-    
+
     sprite->largura_frame = largura_sheet / frames_por_linha;
     int linhas = (frames_total + frames_por_linha - 1) / frames_por_linha;
     sprite->altura_frame = altura_sheet / linhas;
-    
+
     return sprite;
 }
 
 void atualizar_sprite_animado(SpriteAnimado* sprite, float delta_time) {
     if (!sprite || !sprite->carregado) return;
-    
+
     sprite->tempo_acumulado += delta_time;
-    
+
     if (sprite->tempo_acumulado >= sprite->tempo_por_frame) {
         sprite->tempo_acumulado -= sprite->tempo_por_frame;
         sprite->frame_atual++;
-        
+
         if (sprite->frame_atual >= sprite->frames_total) {
             if (sprite->loop) {
                 sprite->frame_atual = 0;
-            } else {
+            }
+            else {
                 sprite->frame_atual = sprite->frames_total - 1;
             }
         }
@@ -145,13 +147,13 @@ void atualizar_sprite_animado(SpriteAnimado* sprite, float delta_time) {
 
 void desenhar_sprite_animado(const SpriteAnimado* sprite, float x, float y, float escala) {
     if (!sprite || !sprite->carregado) return;
-    
+
     int frame_x = (sprite->frame_atual % sprite->frames_por_linha);
     int frame_y = (sprite->frame_atual / sprite->frames_por_linha);
-    
-    float sx = frame_x * sprite->largura_frame;
-    float sy = frame_y * sprite->altura_frame;
-    
+
+    float sx = (float)frame_x * sprite->largura_frame;
+    float sy = (float)frame_y * sprite->altura_frame;
+
     al_draw_scaled_bitmap(
         sprite->spritesheet,
         sx, sy, sprite->largura_frame, sprite->altura_frame,
@@ -163,23 +165,23 @@ void desenhar_sprite_animado(const SpriteAnimado* sprite, float x, float y, floa
     );
 }
 
-void desenhar_sprite_animado_camera(const SpriteAnimado* sprite, 
-                                   float x, float y,
-                                   float camera_x, float camera_y,
-                                   float escala, float zoom) {
+void desenhar_sprite_animado_camera(const SpriteAnimado* sprite,
+    float x, float y,
+    float camera_x, float camera_y,
+    float escala, float zoom) {
     if (!sprite || !sprite->carregado) return;
-    
+
     int frame_x = (sprite->frame_atual % sprite->frames_por_linha);
     int frame_y = (sprite->frame_atual / sprite->frames_por_linha);
-    
-    float sx = frame_x * sprite->largura_frame;
-    float sy = frame_y * sprite->altura_frame;
-    
+
+    float sx = (float)frame_x * sprite->largura_frame;
+    float sy = (float)frame_y * sprite->altura_frame;
+
     float x_tela = (x - camera_x) * zoom;
     float y_tela = (y - camera_y) * zoom;
     float largura_escalada = sprite->largura_frame * escala * zoom;
     float altura_escalada = sprite->altura_frame * escala * zoom;
-    
+
     al_draw_scaled_bitmap(
         sprite->spritesheet,
         sx, sy, sprite->largura_frame, sprite->altura_frame,
@@ -204,7 +206,7 @@ void set_loop_animado(SpriteAnimado* sprite, bool loop) {
 
 void destruir_sprite_animado(SpriteAnimado* sprite) {
     if (!sprite) return;
-    
+
     if (sprite->spritesheet) {
         al_destroy_bitmap(sprite->spritesheet);
     }
@@ -234,16 +236,14 @@ SpriteAnimadoFrameAFrame* criar_sprite_animado_frames(const char* padrao_caminho
     sprite->loop = true;
     sprite->carregado = true;
 
-    // Carrega cada frame usando o padrao (ex: "player_%d.png")
     char caminho_completo[256];
     for (int i = 0; i < frames_total; i++) {
         snprintf(caminho_completo, sizeof(caminho_completo), padrao_caminho, i);
         sprite->frames[i] = al_load_bitmap(caminho_completo);
-        
+
         if (!sprite->frames[i]) {
             fprintf(stderr, "Erro ao carregar frame %d: %s\n", i, caminho_completo);
             sprite->carregado = false;
-            // Libera frames ja carregados
             for (int j = 0; j < i; j++) {
                 if (sprite->frames[j]) al_destroy_bitmap(sprite->frames[j]);
             }
@@ -277,10 +277,9 @@ SpriteAnimadoFrameAFrame* criar_sprite_animado_array(const char** caminhos_arqui
     sprite->loop = true;
     sprite->carregado = true;
 
-    // Carrega cada frame a partir do array de caminhos
     for (int i = 0; i < frames_total; i++) {
         sprite->frames[i] = al_load_bitmap(caminhos_arquivos[i]);
-        
+
         if (!sprite->frames[i]) {
             fprintf(stderr, "Erro ao carregar frame %d: %s\n", i, caminhos_arquivos[i]);
             sprite->carregado = false;
@@ -298,17 +297,18 @@ SpriteAnimadoFrameAFrame* criar_sprite_animado_array(const char** caminhos_arqui
 
 void atualizar_sprite_animado_frames(SpriteAnimadoFrameAFrame* sprite, float delta_time) {
     if (!sprite || !sprite->carregado) return;
-    
+
     sprite->tempo_acumulado += delta_time;
-    
+
     if (sprite->tempo_acumulado >= sprite->tempo_por_frame) {
         sprite->tempo_acumulado -= sprite->tempo_por_frame;
         sprite->frame_atual++;
-        
+
         if (sprite->frame_atual >= sprite->frames_total) {
             if (sprite->loop) {
                 sprite->frame_atual = 0;
-            } else {
+            }
+            else {
                 sprite->frame_atual = sprite->frames_total - 1;
             }
         }
@@ -317,13 +317,13 @@ void atualizar_sprite_animado_frames(SpriteAnimadoFrameAFrame* sprite, float del
 
 void desenhar_sprite_animado_frames(const SpriteAnimadoFrameAFrame* sprite, float x, float y, float escala) {
     if (!sprite || !sprite->carregado) return;
-    
+
     ALLEGRO_BITMAP* frame_atual = sprite->frames[sprite->frame_atual];
     if (!frame_atual) return;
-    
+
     float largura = (float)al_get_bitmap_width(frame_atual);
     float altura = (float)al_get_bitmap_height(frame_atual);
-    
+
     al_draw_scaled_bitmap(
         frame_atual,
         0, 0, largura, altura,
@@ -335,23 +335,23 @@ void desenhar_sprite_animado_frames(const SpriteAnimadoFrameAFrame* sprite, floa
     );
 }
 
-void desenhar_sprite_animado_frames_camera(const SpriteAnimadoFrameAFrame* sprite, 
-                                          float x, float y,
-                                          float camera_x, float camera_y,
-                                          float escala, float zoom) {
+void desenhar_sprite_animado_frames_camera(const SpriteAnimadoFrameAFrame* sprite,
+    float x, float y,
+    float camera_x, float camera_y,
+    float escala, float zoom) {
     if (!sprite || !sprite->carregado) return;
-    
+
     ALLEGRO_BITMAP* frame_atual = sprite->frames[sprite->frame_atual];
     if (!frame_atual) return;
-    
+
     float largura = (float)al_get_bitmap_width(frame_atual);
     float altura = (float)al_get_bitmap_height(frame_atual);
-    
+
     float x_tela = (x - camera_x) * zoom;
     float y_tela = (y - camera_y) * zoom;
     float largura_escalada = largura * escala * zoom;
     float altura_escalada = altura * escala * zoom;
-    
+
     al_draw_scaled_bitmap(
         frame_atual,
         0, 0, largura, altura,
@@ -376,7 +376,7 @@ void set_loop_animacao_frames(SpriteAnimadoFrameAFrame* sprite, bool loop) {
 
 void destruir_sprite_animado_frames(SpriteAnimadoFrameAFrame* sprite) {
     if (!sprite) return;
-    
+
     if (sprite->frames) {
         for (int i = 0; i < sprite->frames_total; i++) {
             if (sprite->frames[i]) {
